pyproject.toml
[tool.poetry]
name = "Mattermost-Printer-Bot"
version = "0.1.0"
description = ""
authors = ["Dirakon"]
# readme = "README.md"
# license = "BSD"
packages = [
  { include = "printer_bot" }
]

[tool.poetry.scripts]
mattermost-printer-bot = "printer_bot.__main__:main"

[tool.poetry.dependencies]
python = "^3.10"
mmpy-bot = "^2.1.4"
httpx = "^0.27.2"
argparse = "^1.4.0"
dateutils = "^0.6.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
